app/requirements.txt
pyspark==3.5.1
pathvalidate==3.2.0
cassandra-driver==3.29.0
tqdm==4.66.2
hdfs==2.7.3